refactor(테스트 코드 패키지 분리) : main 클래스에 대응되도록 test 패키지를 분리 ;
